Add end-to-end test
To avoid installing chrome, this switches our docker container to one purpose-built for this task.
Maybe this image will be generally useful for Angular apps.

Note, I did a bunch of work to use docker-compose with the selenium server in a separate
container. See the add_tests branch. This worked great locally, but the CircleCI support
for docker-compose is super-inconvenient since it doesn't run under the primary container
where we run protractor.